requirements.txt
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
category-encoders==2.6.1
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.5
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
fonttools==4.40.0
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
joblib==1.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.4
llvmlite==0.41.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mizani==0.9.2
mlxtend==0.23.0
nest-asyncio==1.5.8
numba==0.58.1
numpy==1.26.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
patchworklib==0.6.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.11.0
plotnine==0.12.1
plydata==0.4.3
pmdarima==1.8.5
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyclustering==0.10.1.2
pyclustertend @ file:///Users/mi20447/Downloads/pytend/pyclustertend
Pygments==2.16.1
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.1
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.13.0
siuba==0.4.2
six==1.16.0
SQLAlchemy==2.0.16
stack-data==0.6.3
statsmodels==0.14.0
threadpoolctl==3.1.0
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
yellowbrick==1.5
